+++ Assets/lilToon/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dither
 - Alpha replacement for main color 2nd/3rd
 - Property search
+- UV settings to AudioLink mask
 
 ### Fixed
 - Fixed lilToonMulti convert